requirements.txt
pyautogen>=0.2.0
pyaedt>=0.7.0
pydantic>=2.0.0
numpy
pandas
matplotlib
